+++ src/lib.ml
@@ -100,7 +100,7 @@ module Sub_system0 = struct
   module type S = sig
     type t
     type sub_system += T of t
-    val encode : (t -> Syntax.Version.t * Dune_lang.t) option
+    val encode : (t -> Syntax.Version.t * Dune_lang.t list) option
   end
 
   type 'a s = (module S with type t = 'a)
@@ -323,7 +323,7 @@ module Sub_system = struct
       -> lib
       -> Info.t
       -> t
-    val encode : (t -> Syntax.Version.t * Dune_lang.t) option
+    val encode : (t -> Syntax.Version.t * Dune_lang.t list) option
   end
 
   module type S' = sig