+++ src/lib.mli
@@ -312,7 +312,7 @@ module Sub_system : sig
       -> lib
       -> Info.t
       -> t
-    val encode : (t -> Syntax.Version.t * Dune_lang.t) option
+    val encode : (t -> Syntax.Version.t * Dune_lang.t list) option
   end
 
   module Register(M : S) : sig
@@ -320,7 +320,9 @@ module Sub_system : sig
     val get : lib -> M.t option
   end
 
-  val dump_config : lib -> (Syntax.Version.t * Dune_lang.t) Sub_system_name.Map.t
+  val dump_config
+    : lib
+    -> (Syntax.Version.t * Dune_lang.t list) Sub_system_name.Map.t
 end with type lib := t
 
 (** {1 Dependencies for META files} *)